Use u128 for byte sizes
Per issue #58, u64 is insufficient for use with very large sparse files.
Enormous file sizes are also a common filesystem error trope, either
from disk corruption or software bugs, and they're also conceivable with
virtual filesystems.

Handle this as gracefully as can be reasonably expected using 128-bit
integers, which should be sufficient for most uses.

+++ src/interactive/widgets/footer.rs
@@ -11,7 +11,7 @@ use tui::{
 pub struct Footer;
 
 pub struct FooterProps {
-    pub total_bytes: Option<u64>,
+    pub total_bytes: Option<u128>,
     pub entries_traversed: u64,
     pub format: ByteFormat,
     pub message: Option<String>,
